Toggle/Transitions/Audio
Updated sound mixer to include a sub category of system that's called ButtonClick that is being used to lower the sound of the UI buttons to make it less annoying. Applied new category mixer to audio in all buttons. Initiated Progress scene. Updated settings to account for if player hasn't started a game yet.